fix(common): let case conversion pipes accept type unions with `null` (#36259)
The old implementation of case conversion types can handle several
values which are not strings, but the signature did not reflect this.

The new one reports errors when falsy non-string inputs are given to
the pipe (such as `false` or `0`) and has a new signature which
instead reflcts the behaviour on `null` and `undefined`.

Fixes #36259