fix(common): correct and simplify typing of AsyncPipe
`AsyncPipe.transform` will never return `undefined`, even when passed
`undefined` in input, in contrast with what was declared in the
overloads.

Additionally, removing the specialization for `undefined` and `null`
ensures that the result is typed as `unknown | null` by TypeScript,
which can provide compile-time feedback when trying to do a
`valueTypedAsConstantNull | async`.

Finally its signature can be simplified, since regardless of the type
of source, the type of the result is always `T|null`.