fix(common): correct and simplify typing of `KeyValuePipe`
As shown in the tests, `KeyValuePipe.transform` can accept
`undefined`, in which case it always returns `null`.

Additionally, removing the specialization for `null` ensures that when
it is used on a value which is statically known to be `null`, the
result is typed as `unknown | null` by TypeScript, which can provide
useful compile-time feedback.

Finally, put the most generic overload as first, since it is selected
when the input value has type `any`.