fix(common): correct typing and implementation of `SlicePipe`
Even in the overloads, state that it can accept `null` and
`undefined`, in order to ensure easy composition with `async`.

Additionally, change the implementation to return `null` on an
`undefined` input, for consistency with other pipes.

Finally, put the most generic overload as first, since it is selected
when the input value has type `any`.